Refactor operatore, renamed traffic_updater to add_predicted_speeds
