+++ src/mjolnir/CMakeLists.txt
@@ -43,7 +43,7 @@ set(sources
   hierarchybuilder.cc
   ingest_transit.cc
   landmarks.cc
-  traffic_updater.cc
+  add_predicted_speeds.cc
   linkclassification.cc
   luatagtransform.cc
   node_expander.cc
